# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TIC4902 Backend API",
    "version": "1.0.0",
    "description": "REST API for Flink deployment management and authentication. DNS-style names (`deploymentName`, `namespace`) must be lowercase alphanumeric with optional internal hyphens (1–63 chars). Memory fields use a number plus unit, e.g. `1024m`, `2g`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/flink/deployments": {
      "get": {
        "tags": [
          "Flink"
        ],
        "summary": "List deployments",
        "operationId": "listDeployments",
        "responses": {
          "200": {
            "description": "List of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Flink"
        ],
        "summary": "Create deployment",
        "description": "Session mode by default; if `jarId` is set, application mode is used. `namespace`: DNS label pattern, 1–63 chars. `jobParallelism`: 1–1024. `config.jobManager.replicas` must be 1 if set.",
        "operationId": "createDeployment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal or upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "get": {
        "tags": [
          "Flink"
        ],
        "summary": "Get deployment by name",
        "operationId": "getDeployment",
        "responses": {
          "200": {
            "description": "Deployment found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Flink"
        ],
        "summary": "Update deployment",
        "description": "At least one field required. Immutable after create: `namespace`, `deploymentName`, Flink version, `serviceAccount`, `jarId` — use delete + create instead.",
        "operationId": "updateDeployment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deployment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or empty body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Flink"
        ],
        "summary": "Delete deployment",
        "operationId": "deleteDeployment",
        "responses": {
          "200": {
            "description": "Deployment deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/stop": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "post": {
        "tags": ["Flink"],
        "summary": "Gracefully stop a running deployment",
        "description": "Suspends the job after taking a savepoint (upgradeMode `savepoint`). Only valid for application-mode deployments currently in `running` status, and when no other action is already pending. The response returns immediately with `pendingAction: \"stop\"` set; poll GET .../{deploymentName} until the status resolves to `suspended`.",
        "operationId": "stopDeployment",
        "responses": {
          "200": {
            "description": "Stop accepted. Deployment transitions asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "description": "Deployment is session-mode (stop is only supported for application mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Deployment is not currently running, or another action is already pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/force-stop": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "post": {
        "tags": ["Flink"],
        "summary": "Force stop a running deployment",
        "description": "Suspends the job immediately without taking a savepoint (upgradeMode `stateless`). Use only when a clean stop is not possible or state can be discarded. Same preconditions as the graceful stop endpoint.",
        "operationId": "forceStopDeployment",
        "responses": {
          "200": {
            "description": "Force stop accepted. Deployment transitions asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "description": "Deployment is session-mode (force stop is only supported for application mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Deployment is not currently running, or another action is already pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/resume": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "post": {
        "tags": ["Flink"],
        "summary": "Resume a suspended, failed, or unknown-state deployment",
        "description": "Only valid when the deployment is `suspended`, `failed`, or `unknown`, and no other action is already pending. Response returns immediately with `pendingAction: \"resume\"` set; poll GET .../{deploymentName} until status resolves to `running`.",
        "operationId": "resumeDeployment",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resume accepted. Deployment transitions asynchronously",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          },
          "400": {
            "description": "Deployment is session mode, or both savepointId and skipSavepoint were provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found, or savepointId does not belong to this deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Deployment is not suspended/failed/unknown, or another action is already pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/savepoint": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "post": {
        "tags": ["Flink"],
        "summary": "Trigger a manual savepoint",
        "description": "Only valid for a `running` application mode deployment. This call blocks (polling internally, up to 120s) until the savepoint completes, fails, or times out.",
        "operationId": "triggerSavepoint",
        "responses": {
          "200": {
            "description": "Savepoint completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavepointTriggerResponse"
                }
              }  
            }
          },
          "400": {
            "description": "Deployment is session mode or not currently running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Savepoint failed or timed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/savepoints": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "get": {
        "tags": ["Flink"],
        "summary": "List recorded savepoints for a deployment",
        "operationId": "listSavepoints",
        "responses": {
          "200": {
            "description": "Savepoints (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavepointListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flink/deployments/{deploymentName}/diagnostics": {
      "parameters": [
        {
          "name": "deploymentName",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
            "minLength": 1,
            "maxLength": 63
          },
          "description": "Deployment name (DNS subdomain style)"
        }
      ],
      "get": {
        "tags": ["Flink"],
        "summary": "Get deployment diagnostics",
        "description": "Returns Kubernetes-level diagnostics for a Flink deployment, including operator status, conditions, related pods, events, and recommended actions.",
        "operationId": "getDeploymentDiagnostics",
        "responses": {
          "200": {
            "description": "Deployment diagnostics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDiagnostics"
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal or Kubernetes error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Message (500 responses use generic text)"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DnsLabel": {
        "type": "string",
        "pattern": "^[a-z0-9]([-a-z0-9]*[a-z0-9])?$",
        "minLength": 1,
        "maxLength": 63,
        "description": "Lowercase DNS label"
      },
...
      "FlinkConfigCreate": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          },
          "flinkVersion": {
            "type": "string"
          },
          "serviceAccount": {
            "type": "string"
          },
          "jobManager": {
            "$ref": "#/components/schemas/JobManagerSpec"
          },
          "taskManager": {
            "$ref": "#/components/schemas/TaskManagerSpec"
          },
          "flinkConfiguration": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "FlinkConfigUpdate": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string"
          },
          "jobManager": {
            "$ref": "#/components/schemas/JobManagerSpec"
          },
          "taskManager": {
            "$ref": "#/components/schemas/TaskManagerSpec"
          },
          "flinkConfiguration": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DeploymentCreateRequest": {
        "type": "object",
        "required": [
          "deploymentName"
        ],
        "properties": {
          "deploymentName": {
            "$ref": "#/components/schemas/DnsLabel"
          },
          "namespace": {
            "$ref": "#/components/schemas/DnsLabel"
          },
          "jarId": {
            "type": "integer",
            "minimum": 1
          },
          "environmentVariables": {
            "type": "object",
            "additionalProperties": true
          },
          "jobParallelism": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1024
          },
          "config": {
            "$ref": "#/components/schemas/FlinkConfigCreate"
          }
        }
      },
      "DeploymentUpdateRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "environmentVariables": {
            "type": "object",
            "additionalProperties": true
          },
          "jobParallelism": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1024
          },
          "config": {
            "$ref": "#/components/schemas/FlinkConfigUpdate"
          }
        }
      },
      "FlinkDeploymentRef": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "apiVersion": {
            "type": "string"
          }
        }
      },
      "Deployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "deploymentName": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "deploymentMode": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "kubernetesStatus": {
            "type": "object",
            "additionalProperties": true,
            "description": "Live status from Kubernetes operator when available"
          },
          "flinkDeployment": {
            "$ref": "#/components/schemas/FlinkDeploymentRef"
          },
          "environmentVariables": {
            "type": "object",
            "additionalProperties": true
          },
          "jobParallelism": {
            "type": "integer"
          },
          "jar": {
            "type": "object",
            "additionalProperties": true,
            "description": "Present on some create responses when JAR metadata is returned"
          }
        }
      },
...
      "DeploymentListResponse": {
        "type": "object",
        "required": [
          "deployments",
          "total"
        ],
        "properties": {
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Deployment"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "ResumeRequest": {
        "type": "object",
        "description": "At most one of savepointId or skipSavepoint may be set. With neither set, the deployment resumes from its most recent savepoint.",
        "properties": {
          "savepointId": {
            "type": "integer",
            "nullable": true,
            "description": "ID of a previously recorded savepoint (see GET .../savepoints) to resume from"
          },
          "skipSavepoint": {
            "type": "boolean",
            "default": false,
            "description": "Resume without restoring any state (stateless restart)"
          }
        }
      },
      "SavepointTriggerResponse": {
        "type": "object",
        "properties": {
          "savepointPath": {
            "type": "string",
            "description": "Path of the completed savepoint in the state backend"
          }
        }
      },
      "Savepoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "stop"
            ],
            "description": "How the savepoint was created: A manually triggered savepoint, or the automatic one taken on a graceful stop"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SavepointListResponse": {
        "type": "object",
        "properties": {
          "savepoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Savepoint"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "DeploymentDiagnostics": {
        "type": "object",
        "properties": {
          "deploymentName": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeploymentDiagnosticStatus"
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentCondition"
            }
          },
          "pods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentDiagnosticPod"
            }
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentDiagnosticEvent"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentDiagnosticRecommendation"
            }
          }
        }
      },
      "DeploymentDiagnosticStatus": {
        "type": "object",
        "properties": {
          "lifecycleState": {
            "type": "string",
            "nullable": true
          },
          "jobManagerDeploymentStatus": {
            "type": "string",
            "nullable": true
          },
          "jobStatus": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "error": {
            "nullable": true
          }
        }
      },
      "DeploymentCondition": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "lastTransitionTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "DeploymentDiagnosticRecommendation": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "enum": ["low", "medium", "high"]
          },
          "reason": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned from POST /api/auth/login, sent as `Authorization: Bearer <token>`"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}